backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
nltk==3.8.1
python-dotenv==1.0.0
neo4j==5.14.0
requests==2.31.0
